Issue #35 Add Get subject by id

See merge request stefan-j-98/ssa2020-back-hypnotized-chicken!35

+++ SSA2020-Back-Hypnotized-Chicken/SSA2020-Back-Hypnotized-Chicken.API/Controllers/SemestersController.cs
@@ -31,7 +31,7 @@ namespace SSA2020_Back_Hypnotized_Chicken.API.Controllers
             return Ok(semesterMapResult);
         }
         [HttpGet("{id}")]
-        public SemesterDTO GetById(int id)
+        public SemesterDTO GetById(short id)
         {
             var semester = UnitOfWork.SemestersRepository.GetSemesterById(id);
